Reduce test hardening to focused regressions and minimal fixes
