+++ pyproject.toml
@@ -66,7 +66,6 @@ extend-ignore = ["E203", "W503"]
 [tool.pytest.ini_options]
 addopts = "-ra"
 testpaths = ["tests"]
-markers = ["integration: real model, numbering, or serialization pipeline"]
 
 [tool.coverage.run]
 branch = true
