+++ tests/conftest.py
@@ -1,4 +1,4 @@
-import functools
+from pathlib import Path
 
 import pytest
 from Bio.PDB import PDBParser
@@ -6,29 +6,30 @@ from Bio.PDB import PDBParser
 from sabr.alignment import align
 from sabr.model import encode
 from sabr.structure import extract_chain
-from tests.helpers import CASES, DATA
+
+DATA = Path(__file__).parent / "data"
 
 
 @pytest.fixture(scope="session")
-def aligned_case():
-    @functools.cache
-    def prepare(chain_type, mode="sabr"):
-        filename, chain = CASES[chain_type]
-        path = DATA / filename
-        structure = PDBParser(QUIET=True).get_structure(chain_type, path)
+def aligned_cases():
+    parser = PDBParser(QUIET=True)
+    cases = {
+        "H": (DATA / "test_heavy_chain.pdb", "F"),
+        "K": (DATA / "12e8_L.pdb", "L"),
+        "L": (DATA / "1bjm_A.pdb", "A"),
+    }
+    results = {}
+    for expected, (path, chain) in cases.items():
+        structure = parser.get_structure(expected, path)
         data = extract_chain(structure, chain, None)
-        embeddings = encode(data.coords, mode)
         alignment, selected, score = align(
-            embeddings, data.gap_indices, "auto", 0.0, mode=mode
+            encode(data.coords), data.gap_indices, "auto", 0.0
         )
-        return {
+        results[expected] = {
             "alignment": alignment,
-            "embeddings": embeddings,
             "chain": chain,
             "data": data,
             "path": path,
             "selected": selected,
-            "score": score,
         }
-
-    return prepare
+    return results
