--- tests/data/regression_provenance.json
@@ -1,56 +1,0 @@
-{
-  "producing_commit": "2227a34a9ccc93157cb0fc3615b50e673c8b309f",
-  "python": "3.12.14",
-  "platform": "macOS-26.5.2-arm64-arm-64bit",
-  "dependencies": {
-    "biopython": "1.87",
-    "jax": "0.10.2",
-    "jaxlib": "0.10.2",
-    "numpy": "2.5.1",
-    "dm-haiku": "0.0.16"
-  },
-  "backend": "cpu",
-  "parameters": {
-    "noise_level": 0.0,
-    "scfv_schemes": [
-      "imgt"
-    ]
-  },
-  "fixtures": {
-    "8dy0_A.cif": {
-      "source": "https://files.rcsb.org/download/8DY0.cif",
-      "entry": "8DY0",
-      "author_chain": "A",
-      "source_sha256": "218677ee086b4fda42004028a41e5fff11420546ede481b8a957737f18f411c5",
-      "extracted_sha256": "26a3dd1f09b4eb3ebcf098130e04a6fd77338a47c217bc5785ff5512366f1a33",
-      "extraction": "Biopython MMCIFIO, first/only model, selected author chain including hetero residues and all atom conformers; atomic categories only.",
-      "polymer_residues": 244,
-      "gap_indices": []
-    },
-    "8dy1_C.cif": {
-      "source": "https://files.rcsb.org/download/8DY1.cif",
-      "entry": "8DY1",
-      "author_chain": "C",
-      "source_sha256": "9f47dc11efce22ebf2d71f08e396ba3c59f4dcf0a0bf1580f0d314006a82b4b2",
-      "extracted_sha256": "1eb3631db1441231c0f782d001115b444ef40f2ebafeb42abdf11fb6345ff0ef",
-      "extraction": "Biopython MMCIFIO, first/only model, selected author chain including hetero residues and all atom conformers; atomic categories only.",
-      "polymer_residues": 228,
-      "gap_indices": [
-        110
-      ]
-    }
-  },
-  "inputs": {
-    "test_heavy_chain.pdb": "a7c67fd00db4e882737ca3e91266ca729af47c9f95ada789b44cc0b8e617d696",
-    "12e8_L.pdb": "e539d09539f260b1e2795d1298e304bb30e5ecd1d0da434bd5474664ed1ebc6d",
-    "1bjm_A.pdb": "06d123297f10bcc1bfb998da3c98b7623849fe930acff21bbee042b231297f68"
-  },
-  "review": "Reviewed query-row/sequence parity, complete residue counts, unique IDs, domain boundaries and mode differences against unchanged commit 2227a34. These are regression captures, not independent biological accuracy labels. See tests/data/README.md for the SoftAlign linker limitation.",
-  "outputs": {
-    "8dy0_A.cif": "26a3dd1f09b4eb3ebcf098130e04a6fd77338a47c217bc5785ff5512366f1a33",
-    "8dy1_C.cif": "1eb3631db1441231c0f782d001115b444ef40f2ebafeb42abdf11fb6345ff0ef",
-    "multidomain_baseline.json": "aa8e48d29a5068153c6d2756a0650d7eaa115edf23d506b1e2f936b24efb8605",
-    "softalign_math_baseline.npz": "0a3c4a89036e15409516f7ff03956c61ad716ee7587a1ba8858f623f2d10db33",
-    "softalign_numbering_baseline.json": "7751ebff5a95d108111f3b73b8e722719593be276e3ff0ceb2d262dcab955ca3"
-  }
-}
