+++ tests/test_api.py
@@ -8,6 +8,7 @@ from Bio.PDB.Structure import Structure
 
 import sabr.api as api
 from sabr import constants, renumber_structure
+from sabr.structure import extract_chain
 
 DATA = Path(__file__).parent / "data"
 
@@ -64,6 +65,8 @@ def test_biopython_structure_is_copied_before_numbering(monkeypatch):
     path = DATA / "test_heavy_chain.pdb"
     structure = PDBParser(QUIET=True).get_structure("structure", path)
     original_ids = _bio_ids(structure, "F")
+    structure.header["nested"] = ["original"]
+    original_coords = np.array([atom.coord for atom in structure.get_atoms()])
 
     result = renumber_structure(structure, "F", chain_type="H")
 
@@ -71,6 +74,18 @@ def test_biopython_structure_is_copied_before_numbering(monkeypatch):
     assert result is not structure
     assert _bio_ids(structure, "F") == original_ids
     assert _bio_ids(result, "F") != original_ids
+    np.testing.assert_array_equal(
+        [atom.coord for atom in result.get_atoms()], original_coords
+    )
+    assert [r.resname for r in result.get_residues()] == [
+        r.resname for r in structure.get_residues()
+    ]
+    next(result.get_atoms()).coord[0] += 100
+    result.header["nested"].append("changed")
+    np.testing.assert_array_equal(
+        [atom.coord for atom in structure.get_atoms()], original_coords
+    )
+    assert structure.header["nested"] == ["original"]
 
 
 def test_non_target_content_and_out_of_range_residues_are_preserved(
@@ -111,6 +126,14 @@ def test_non_target_content_and_out_of_range_residues_are_preserved(
     assert result[0]["Z"][other_residue.id]["O"].coord.tolist() == (
         original_other[other_residue.id]["O"].coord.tolist()
     )
+
+
+def test_numeric_residue_range_includes_insertion_codes():
+    structure = PDBParser(QUIET=True).get_structure(
+        "insertions", DATA / "test_insertion_codes.pdb"
+    )
+    data = extract_chain(structure, "A", (52, 52))
+    assert data.residue_ids == ((52, ""), (52, "A"), (52, "B"))
 
 
 def test_range_that_would_collide_with_unchanged_ids_is_rejected(
@@ -285,6 +308,47 @@ def test_tcr_chain_uses_k_reference_and_actual_numbering_type(
     }
 
 
+def test_missing_backbone_is_reported_with_the_residue():
+    structure = PDBParser(QUIET=True).get_structure(
+        "missing", DATA / "test_no_backbone.pdb"
+    )
+    with pytest.raises(ValueError, match="missing required backbone"):
+        renumber_structure(structure, "A")
+
+
+@pytest.mark.parametrize(
+    ("chain", "message"),
+    [("", "non-empty"), ("missing", "not found")],
+)
+def test_invalid_or_missing_chain_is_reported(chain, message):
+    structure = PDBParser(QUIET=True).get_structure(
+        "heavy", DATA / "test_heavy_chain.pdb"
+    )
+    with pytest.raises(ValueError, match=message):
+        renumber_structure(structure, chain)
+
+
+def test_missing_backbone_outside_selected_range_is_ignored():
+    structure = PDBParser(QUIET=True).get_structure(
+        "heavy", DATA / "test_heavy_chain.pdb"
+    )
+    list(structure[0]["F"])[-1].detach_child("N")
+    data = extract_chain(structure, "F", (2, 127))
+    assert data.residue_ids[0] == (2, "")
+    assert data.residue_ids[-1] == (127, "")
+
+
+def test_multiple_models_are_rejected():
+    structure = PDBParser(QUIET=True).get_structure(
+        "heavy", DATA / "test_heavy_chain.pdb"
+    )
+    second = copy.deepcopy(structure[0])
+    second.id = 1
+    structure.add(second)
+    with pytest.raises(ValueError, match="exactly one model"):
+        renumber_structure(structure, "F")
+
+
 @pytest.mark.parametrize(
     ("kwargs", "message"),
     [
@@ -316,20 +380,23 @@ def test_invalid_public_options_fail_before_model_execution(kwargs, message):
         renumber_structure(structure, "F", **kwargs)
 
 
+def test_unsupported_structure_type_raises_type_error():
+    with pytest.raises(TypeError, match="Bio.PDB"):
+        renumber_structure(object(), "A")
+
+
 @pytest.mark.parametrize("chain_type", constants.TCR_CHAIN_TYPES)
 @pytest.mark.parametrize("scheme", ("chothia", "kabat", "martin", "wolfguy"))
-def test_tcr_invalid_scheme_fails_before_encoding(
+def test_invalid_tcr_scheme_fails_before_encoding(
     monkeypatch, chain_type, scheme
 ):
     monkeypatch.setattr(
         api,
         "encode",
-        lambda *args, **kwargs: pytest.fail(
-            "unsupported TCR scheme reached encoder"
-        ),
+        lambda *args: pytest.fail("unsupported TCR scheme reached encoder"),
     )
     structure = PDBParser(QUIET=True).get_structure(
-        "input", DATA / "test_heavy_chain.pdb"
+        "heavy", DATA / "test_heavy_chain.pdb"
     )
     with pytest.raises(ValueError, match="only IMGT or AHo"):
         renumber_structure(structure, "F", chain_type=chain_type, scheme=scheme)
