+++ tests/test_cli.py
@@ -1,10 +1,12 @@
+import json
 import logging
 import shutil
 from pathlib import Path
 
 import click
+import numpy as np
 import pytest
-from Bio.PDB import MMCIFParser, PDBParser
+from Bio.PDB import PDBIO, MMCIFParser, PDBParser
 from click.testing import CliRunner
 
 import sabr.cli as cli
@@ -82,6 +84,11 @@ def test_cli_maps_the_complete_compact_interface(monkeypatch, tmp_path):
 
 
 def test_cli_defaults_are_deterministic_and_quiet(monkeypatch, tmp_path):
+    monkeypatch.setattr(
+        cli.jax,
+        "default_backend",
+        lambda: pytest.fail("backend initialized for suppressed logging"),
+    )
     captured = {}
     monkeypatch.setattr(cli, "renumber_structure", _passthrough(captured))
     output = tmp_path / "numbered.pdb"
@@ -311,6 +318,58 @@ def test_cli_failure_leaves_no_output_or_temporary_file(monkeypatch, tmp_path):
     assert list(tmp_path.iterdir()) == []
 
 
+@pytest.mark.parametrize("mode", ("sabr", "softalign"))
+def test_real_cli_round_trip(tmp_path, mode):
+    source = PDBParser(QUIET=True).get_structure(
+        "input", DATA / "test_heavy_chain.pdb"
+    )
+    # The fixture is already IMGT numbered. Offset IDs so a no-op cannot pass.
+    for residue in source[0]["F"]:
+        residue.id = (residue.id[0], residue.id[1] + 500, residue.id[2])
+    input_path = tmp_path / "offset.pdb"
+    writer = PDBIO()
+    writer.set_structure(source)
+    writer.save(str(input_path))
+    output = tmp_path / "numbered.pdb"
+    result = CliRunner().invoke(
+        cli.main,
+        [
+            "-i",
+            str(input_path),
+            "-c",
+            "F",
+            "-o",
+            str(output),
+            "-t",
+            "H,K",
+            "--mode",
+            mode,
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    parsed = PDBParser(QUIET=True).get_structure("output", output)
+    golden = json.loads((DATA / "numbering_baseline.json").read_text())["H"][
+        "schemes"
+    ]["imgt"]
+    assert [(r.id[1], r.id[2].strip()) for r in parsed[0]["F"]] == [
+        (n, code.strip()) for n, code, _ in golden["numbered"]
+    ]
+    assert [r.resname for r in parsed.get_residues()] == [
+        r.resname for r in source.get_residues()
+    ]
+    before, after = list(source.get_atoms()), list(parsed.get_atoms())
+    assert len(before) == len(after)
+    for a, b in zip(before, after, strict=True):
+        assert (a.name, a.altloc, a.element, a.occupancy, a.bfactor) == (
+            b.name,
+            b.altloc,
+            b.element,
+            b.occupancy,
+            b.bfactor,
+        )
+        np.testing.assert_allclose(a.coord, b.coord, rtol=0, atol=0.00051)
+
+
 def test_cli_rejects_structural_gap_without_override(tmp_path):
     output = tmp_path / "8sve-numbered.cif"
     result = CliRunner().invoke(
@@ -328,6 +387,37 @@ def test_cli_rejects_structural_gap_without_override(tmp_path):
     assert "SAbR will not run" in result.output
     assert "--dangerously-allow-structural-gaps" in result.output
     assert not output.exists()
+
+
+def test_8sve_cli_mmcif_round_trip_matches_full_golden(tmp_path):
+    golden = json.loads((DATA / "8sve_cdr1_imgt.json").read_text())
+    output = tmp_path / "8sve-numbered.cif"
+    result = CliRunner().invoke(
+        cli.main,
+        [
+            "-i",
+            str(DATA / "8sve_L.pdb"),
+            "-c",
+            "M",
+            "-o",
+            str(output),
+            "-t",
+            "K",
+            "--dangerously-allow-structural-gaps",
+        ],
+    )
+    assert result.exit_code == 0, result.output
+    assert result.output.startswith(
+        "WARNING: Structural-gap safety check disabled by "
+        "--dangerously-allow-structural-gaps"
+    )
+    parsed = MMCIFParser(QUIET=True).get_structure("output", output)
+    actual = [
+        [residue.id[1], residue.id[2].strip()]
+        for residue in parsed[0]["M"]
+        if not residue.id[0].strip()
+    ]
+    assert actual == [[item[2], item[3]] for item in golden["mapping"]]
 
 
 def test_cli_accepts_mmcif_input(monkeypatch, tmp_path):
@@ -357,7 +447,13 @@ def test_cli_accepts_mmcif_input(monkeypatch, tmp_path):
 def test_cli_rejects_unsupported_file_extensions(
     monkeypatch, tmp_path, invalid_side
 ):
-    monkeypatch.setattr(cli, "renumber_structure", _passthrough())
+    monkeypatch.setattr(
+        cli,
+        "renumber_structure",
+        lambda *args, **kwargs: pytest.fail(
+            "unsupported extension reached pipeline"
+        ),
+    )
     input_path = DATA / "test_heavy_chain.pdb"
     output_path = tmp_path / "numbered.pdb"
     if invalid_side == "input":
@@ -374,15 +470,28 @@ def test_cli_rejects_unsupported_file_extensions(
     assert not output_path.exists()
 
 
-def test_writer_failure_removes_partial_temporary_file(monkeypatch, tmp_path):
+@pytest.mark.parametrize(
+    "stage,error",
+    [("write", OSError), ("write", click.ClickException), ("replace", OSError)],
+)
+@pytest.mark.parametrize("existing", [False, True])
+def test_writer_failure_removes_partial_temporary_file(
+    monkeypatch, tmp_path, stage, error, existing
+):
     monkeypatch.setattr(cli, "renumber_structure", _passthrough())
 
     def fail_after_partial_write(structure, path):
-        path.write_text("partial")
-        raise OSError("planned writer failure")
+        if stage == "write":
+            path.write_text("partial")
+        raise error("planned writer failure")
 
-    monkeypatch.setattr(cli, "_write_structure", fail_after_partial_write)
+    if stage == "write":
+        monkeypatch.setattr(cli, "_write_structure", fail_after_partial_write)
+    else:
+        monkeypatch.setattr(cli.os, "replace", fail_after_partial_write)
     output = tmp_path / "failed.pdb"
+    if existing:
+        output.write_bytes(b"original")
     result = CliRunner().invoke(
         cli.main,
         [
@@ -392,11 +501,14 @@ def test_writer_failure_removes_partial_temporary_file(monkeypatch, tmp_path):
             "F",
             "-o",
             str(output),
+            "--overwrite",
         ],
     )
     assert result.exit_code != 0
     assert "planned writer failure" in result.output
-    assert list(tmp_path.iterdir()) == []
+    assert list(tmp_path.iterdir()) == ([output] if existing else [])
+    if existing:
+        assert output.read_bytes() == b"original"
 
 
 def test_cleanup_failure_does_not_replace_original_error(monkeypatch, tmp_path):
@@ -515,122 +627,3 @@ def test_help_and_version_are_available():
     assert "--no-mmcif" in help_result.output
     assert "--dangerously-allow-structural-gaps" in help_result.output
     assert version_result.exit_code == 0
-
-
-@pytest.mark.parametrize("failure", ("writer", "replace", "click"))
-def test_failed_overwrite_preserves_destination_and_cleans_temporary(
-    monkeypatch, tmp_path, failure
-):
-    monkeypatch.setattr(cli, "renumber_structure", _passthrough())
-    output = tmp_path / "existing.pdb"
-    output.write_bytes(b"original destination\n")
-    if failure in ("writer", "click"):
-
-        def fail(structure, path):
-            path.write_text("partial")
-            if failure == "click":
-                raise click.ClickException("planned click failure")
-            raise OSError("planned writer failure")
-
-        monkeypatch.setattr(cli, "_write_structure", fail)
-    else:
-
-        def fail(source, destination):
-            assert source.exists()
-            raise OSError("planned replace failure")
-
-        monkeypatch.setattr(cli.os, "replace", fail)
-    result = CliRunner().invoke(
-        cli.main,
-        [
-            "-i",
-            str(DATA / "test_heavy_chain.pdb"),
-            "-c",
-            "F",
-            "-o",
-            str(output),
-            "--overwrite",
-        ],
-    )
-    assert result.exit_code != 0
-    assert f"planned {failure} failure" in result.output
-    assert output.read_bytes() == b"original destination\n"
-    assert list(tmp_path.iterdir()) == [output]
-
-
-@pytest.mark.parametrize("automatic_cif", (False, True))
-def test_successful_overwrite_replaces_the_resolved_destination(
-    monkeypatch, tmp_path, automatic_cif
-):
-    fake = _with_extended_insertion_code() if automatic_cif else _passthrough()
-    monkeypatch.setattr(cli, "renumber_structure", fake)
-    requested = tmp_path / "numbered.pdb"
-    destination = requested.with_suffix(".cif") if automatic_cif else requested
-    destination.write_bytes(b"old")
-    if automatic_cif:
-        requested.write_bytes(b"keep requested PDB")
-    result = CliRunner().invoke(
-        cli.main,
-        [
-            "-i",
-            str(DATA / "test_heavy_chain.pdb"),
-            "-c",
-            "F",
-            "-o",
-            str(requested),
-            "--overwrite",
-        ],
-    )
-    assert result.exit_code == 0, result.output
-    parser = MMCIFParser(QUIET=True) if automatic_cif else PDBParser(QUIET=True)
-    parsed = parser.get_structure("output", destination)
-    assert len(parsed[0]["F"]) > 1
-    assert next(parsed.get_residues()).id[2] == ("AA" if automatic_cif else " ")
-    if automatic_cif:
-        assert requested.read_bytes() == b"keep requested PDB"
-    assert not list(tmp_path.glob(".*.tmp*"))
-
-
-def test_invalid_output_extension_fails_before_pipeline(monkeypatch, tmp_path):
-    monkeypatch.setattr(
-        cli,
-        "renumber_structure",
-        lambda *args, **kwargs: pytest.fail("invalid output reached pipeline"),
-    )
-    result = CliRunner().invoke(
-        cli.main,
-        [
-            "-i",
-            str(DATA / "test_heavy_chain.pdb"),
-            "-c",
-            "F",
-            "-o",
-            str(tmp_path / "bad.txt"),
-        ],
-    )
-    assert result.exit_code != 0
-    assert "Output must use" in result.output
-    assert not list(tmp_path.iterdir())
-
-
-def test_quiet_cli_does_not_initialize_backend_for_logging(
-    monkeypatch, tmp_path
-):
-    monkeypatch.setattr(cli, "renumber_structure", _passthrough())
-    monkeypatch.setattr(
-        cli.jax,
-        "default_backend",
-        lambda: pytest.fail("backend initialized for suppressed logging"),
-    )
-    result = CliRunner().invoke(
-        cli.main,
-        [
-            "-i",
-            str(DATA / "test_heavy_chain.pdb"),
-            "-c",
-            "F",
-            "-o",
-            str(tmp_path / "numbered.pdb"),
-        ],
-    )
-    assert result.exit_code == 0, result.output
